# openapi/personnel-administration-api.yaml
openapi: 3.0.3
info:
  title: B4RRHH Personnel Administration API
  version: 0.1.0
...
paths:
...
  /employees/{ruleSystemCode}/{employeeTypeCode}/{employeeNumber}/contacts:
    post:
      operationId: createContactByBusinessKey
      summary: Create employee contact
      tags:
        - Employee Contact
      parameters:
        - $ref: "#/components/parameters/AcceptLanguage"
        - in: path
          name: ruleSystemCode
          required: true
          schema:
            type: string
            maxLength: 5
        - in: path
          name: employeeTypeCode
          required: true
          schema:
            type: string
            maxLength: 30
        - in: path
          name: employeeNumber
          required: true
          schema:
            type: string
            maxLength: 15
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContactRequest"
      responses:
        "201":
          description: Contact created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactResponse"
        "400":
          description: Invalid request
        "404":
          description: Rule system, employee, or catalog not found
        "409":
          description: Contact already exists for employee and contact type
    get:
      operationId: listEmployeeContactsByBusinessKey
      summary: List employee contacts
      tags:
        - Employee Contact
      parameters:
        - $ref: "#/components/parameters/AcceptLanguage"
        - in: path
          name: ruleSystemCode
          required: true
          schema:
            type: string
            maxLength: 5
        - in: path
          name: employeeTypeCode
          required: true
          schema:
            type: string
            maxLength: 30
        - in: path
          name: employeeNumber
          required: true
          schema:
            type: string
            maxLength: 15
      responses:
        "200":
          description: Contact list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContactResponse"
        "404":
          description: Rule system or employee not found
  /employees/{ruleSystemCode}/{employeeTypeCode}/{employeeNumber}/contacts/{contactTypeCode}:
    get:
      operationId: getContactByBusinessKey
      summary: Get employee contact by business key
      tags:
        - Employee Contact
      parameters:
        - $ref: "#/components/parameters/AcceptLanguage"
        - in: path
          name: ruleSystemCode
          required: true
          schema:
            type: string
            maxLength: 5
        - in: path
          name: employeeTypeCode
          required: true
          schema:
            type: string
            maxLength: 30
        - in: path
          name: employeeNumber
          required: true
          schema:
            type: string
            maxLength: 15
        - in: path
          name: contactTypeCode
          required: true
          schema:
            type: string
            maxLength: 30
      responses:
        "200":
          description: Contact found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactResponse"
        "404":
          description: Rule system, employee, or contact not found
    put:
      operationId: updateContactByBusinessKey
      summary: Update employee contact
      tags:
        - Employee Contact
      parameters:
        - $ref: "#/components/parameters/AcceptLanguage"
        - in: path
          name: ruleSystemCode
          required: true
          schema:
            type: string
            maxLength: 5
        - in: path
          name: employeeTypeCode
          required: true
          schema:
            type: string
            maxLength: 30
        - in: path
          name: employeeNumber
          required: true
          schema:
            type: string
            maxLength: 15
        - in: path
          name: contactTypeCode
          required: true
          schema:
            type: string
            maxLength: 30
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateContactRequest"
      responses:
        "200":
          description: Contact updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactResponse"
        "400":
          description: Invalid request
        "404":
          description: Rule system, employee, catalog, or contact not found
    delete:
      operationId: deleteContactByBusinessKey
      summary: Delete employee contact
      tags:
        - Employee Contact
      parameters:
        - in: path
          name: ruleSystemCode
          required: true
          schema:
            type: string
            maxLength: 5
        - in: path
          name: employeeTypeCode
          required: true
          schema:
            type: string
            maxLength: 30
        - in: path
          name: employeeNumber
          required: true
          schema:
            type: string
            maxLength: 15
        - in: path
          name: contactTypeCode
          required: true
          schema:
            type: string
            maxLength: 30
      responses:
        "204":
          description: Contact deleted
        "404":
          description: Rule system, employee, or contact not found
...
  /work-centers/{ruleSystemCode}/{workCenterCode}/contacts/{contactNumber}:
    get:
      operationId: getWorkCenterContact
      summary: Get work center contact by business key
      tags:
        - Work Center Contact
      parameters:
        - in: path
          name: ruleSystemCode
          required: true
          schema:
            type: string
            maxLength: 5
        - in: path
          name: workCenterCode
          required: true
          schema:
            type: string
            maxLength: 30
        - in: path
          name: contactNumber
          required: true
          schema:
            type: integer
            format: int32
      responses:
        "200":
          description: Work center contact found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkCenterContactResponse"
        "400":
          description: Invalid path parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkCenterErrorResponse"
        "404":
          description: Work center or contact does not exist for the business key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkCenterErrorResponse"
        "409":
          description: Work center exists but is not applicable today
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkCenterErrorResponse"
    put:
      operationId: updateWorkCenterContact
      summary: Update work center contact by business key
      tags:
        - Work Center Contact
      parameters:
        - in: path
          name: ruleSystemCode
          required: true
          schema:
            type: string
            maxLength: 5
        - in: path
          name: workCenterCode
          required: true
          schema:
            type: string
            maxLength: 30
        - in: path
          name: contactNumber
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWorkCenterContactRequest"
      responses:
        "200":
          description: Work center contact updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkCenterContactResponse"
        "400":
          description: Invalid request payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkCenterErrorResponse"
        "404":
          description: Work center or contact does not exist for the business key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkCenterErrorResponse"
        "409":
          description: Work center exists but is not applicable today
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkCenterErrorResponse"
    delete:
      operationId: deleteWorkCenterContact
      summary: Delete work center contact by business key
      tags:
        - Work Center Contact
      parameters:
        - in: path
          name: ruleSystemCode
          required: true
          schema:
            type: string
            maxLength: 5
        - in: path
          name: workCenterCode
          required: true
          schema:
            type: string
            maxLength: 30
        - in: path
          name: contactNumber
          required: true
          schema:
            type: integer
            format: int32
      responses:
        "204":
          description: Work center contact deleted
        "400":
          description: Invalid path parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkCenterErrorResponse"
        "404":
          description: Work center or contact does not exist for the business key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkCenterErrorResponse"
        "409":
          description: Work center exists but is not applicable today
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkCenterErrorResponse"
...
components:
  parameters:
    AcceptLanguage:
      in: header
      name: Accept-Language
      required: false
      description: >-
        Language of the catalog labels in the response (ADR-052). Short BCP 47 tag such as
        `es-ES`, `fr-FR` or `en`; the heaviest language of the header is used. Codes never
        change. A label without a translation for that language, a missing header or an
        unrecognised one all fall back to the base literal, silently.
      schema:
        type: string
        example: es-ES
  schemas:
...
    CreateContactRequest:
      type: object
      additionalProperties: false
      required:
        - contactTypeCode
        - contactValue
      properties:
        contactTypeCode:
          type: string
          maxLength: 30
        contactValue:
          type: string
          maxLength: 300
    UpdateContactRequest:
      type: object
      additionalProperties: false
      required:
        - contactValue
      properties:
        contactValue:
          type: string
          maxLength: 300
    ContactResponse:
      type: object
      required:
        - contactTypeCode
        - contactValue
      properties:
        contactTypeCode:
          type: string
        contactTypeName:
          type: string
          nullable: true
          description: Optional display name from CONTACT_TYPE rule entities in the employee ruleSystemCode catalog.
        contactValue:
          type: string
...
    WorkCenterContactResponse:
      type: object
      required:
        - contactNumber
        - contactTypeCode
        - contactValue
      properties:
        contactNumber:
          type: integer
          format: int32
        contactTypeCode:
          type: string
          maxLength: 30
        contactTypeName:
          type: string
          nullable: true
          description: Optional display name from CONTACT_TYPE rule entities in the work center ruleSystemCode catalog.
        contactValue:
          type: string
          maxLength: 300
...
    UpdateWorkCenterContactRequest:
      type: object
      required:
        - contactTypeCode
        - contactValue
      properties:
        contactTypeCode:
          type: string
          maxLength: 30
        contactValue:
          type: string
          maxLength: 300
...
    WorkCenterErrorResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
...